Don't set extension data for implicit isin
Generate implicit isin on the fly.  In the next patch in this chain, we
switch to using a page property for the data and storing the parent for
each content subpage on a wiki doesn't seem great.  Also, this avoid the
need to cast Parsoid's CMC to a ParserOutput to use a setter.

Bug: T389523
Change-Id: I200fa13ef85ff103e5c69e8e5a2b6c2872a91933