Improve GeoCrumbs fallback when page property is not (yet) set
During the transition period, page properties will not yet be
set in the database (because RefreshLinksJob has not run on
every page, and that job is typically only run on an edit)
but on a cache miss, reparsing the page won't put the is-in
in the extension data any more.

Workaround this by also checking the page properties of the
reparsed page.

Bug: T391128
Change-Id: Ic0dd05ee30a64d2318462b35722d2e21864cc438
(cherry picked from commit 95dbbc6072b1413781a77ce293b5de44ba1d4dcc)